почти доделал offer

+++ src/wish/wish.service.ts
@@ -1,5 +1,5 @@
 import { User } from './../users/entities/user.entity';
-import { Injectable } from '@nestjs/common';
+import { ConflictException, Injectable } from '@nestjs/common';
 import { CreateWishDto } from './dto/create-wish.dto';
 import { UpdateWishDto } from './dto/update-wish.dto';
 import { FindManyOptions, Like, Repository } from 'typeorm';
@@ -22,14 +22,16 @@ export class WishService {
     return await this.wishRepository.save(wish);
   }
 
-  async createOffer(createWishDto: CreateWishDto, offer: Offer) {
-    const wish = this.wishRepository.create({
+  async createPriceWish(createWishDto: CreateWishDto, amount: number, offer: Offer, id: number) {
+    const wish = await this.findOne(id)
+    const updateWish = this.wishRepository.create({
       ...createWishDto,
-      offers: [offer],
-      rised: 1
+      offers: [...wish.offers, offer],
+      raised: wish.raised + amount
     })
-    return await this.wishRepository.save(wish)
+    return await this.wishRepository.save(updateWish)
   }
+
 
   async findAll() {
     const wishes = await this.wishRepository.find();
@@ -48,7 +50,8 @@ export class WishService {
     const wish = await this.wishRepository.findOne({
       where: { id: wishId },
       relations: {
-        owner: true
+        owner: true,
+        offers: true
       }
     });
     return wish
@@ -67,11 +70,12 @@ export class WishService {
 
   async copyWishById(wishId: number, user: User) {
     const wish = await this.findOne(wishId);
-    const copyWish = this.wishRepository.create({ ...wish, owner: user });
-    const originalWish = this.wishRepository.create({ ...wish, copied: wish.copied + 1 });
+    const copyWish = this.wishRepository.create({ ...wish, owner: user, raised: 0  });
+    const originalWish = this.wishRepository.create({ ...wish, copied: wish.copied + 1});
     await this.wishRepository.insert(copyWish);
     return await this.wishRepository.save(originalWish); 
   }
+
 
   async wishAllByUsername( username: string ): Promise<Wish[]> {
     const wish = await this.wishRepository.find({
@@ -101,7 +105,11 @@ export class WishService {
   }
 
   async remove(id: number) {
-    const wish = await this.findOne(id)
-    return await this.wishRepository.remove(wish)
+    const wish = await this.findOne(id);
+    if (wish.raised === 0) {
+      return await this.wishRepository.remove(wish)
+    } else {
+      throw new ConflictException('rised is not null');
+    }
   }
 }
